Use one initrd on Clear Linux
This is a special patch only used by Clear Linux, not to be upstream.

Clear Linux is shipping an unique initrd image which will be used by all
the kernels. This initrd is used to decrypt the root partition.

The initrd is called clr-init.img.gz which is shipped in the
bootloader-extra bundle. When this bundle is installed clr-boot-manager
will install the initrd image when the update operation is performed.