Add temporary workaround to remove kvm kernel
The kvm kernel has been obsoleted and should be cleaned up. For CBM
this means adding logic to no longer allow the default kernel type to
be kvm and that even the latest release of the kvm type should be
removed. Once a format bump has happened this patch can be removed.

Signed-off-by: William Douglas <william.douglas@intel.com>